liang

+++ bester-api/src/main/java/com/xianbester/api/dto/OrderRecordCountDTO.java
@@ -12,7 +12,7 @@ import java.math.BigDecimal;
 public class OrderRecordCountDTO implements Serializable {
     private static final long serialVersionUID = -7238780244745508896L;
     private BigDecimal price;
-    private BigDecimal frequency;
-    private BigDecimal peopleNum;
+    private Long frequency;
+    private Long peopleNum;
     private BigDecimal averagePrice;
 }
